ffmpeg perspective command practice
you can plug in the x and y coordinates for each corner to see how the perspective of your video command will change. Change the data in the cells with the gray background. Copy the red text into your terminal emulator and press enter. then your video should start to play with the new perspective that you have defined.